[dynamo] Organize higherorderops variable trackers
[ghstack-poisoned]
