+++ torch/_dynamo/utils.py
@@ -1703,7 +1703,7 @@ def is_utils_checkpoint(obj):
 
 def build_checkpoint_variable(**options):
     import torch._higher_order_ops.wrap as higher_order_ops
-    from .variables.torch import TorchHigherOrderOperatorVariable
+    from .variables.higher_order_ops import TorchHigherOrderOperatorVariable
 
     # TODO - This is a temporary sitaution where we have two versions of
     # checkpointing implemetation. We will converge on one and remove the other.
@@ -1711,7 +1711,7 @@ def build_checkpoint_variable(**options):
     if torch._functorch.config.functionalize_rng_ops:
         activation_checkpoint_op = higher_order_ops.wrap_activation_checkpoint
 
-    return TorchHigherOrderOperatorVariable(
+    return TorchHigherOrderOperatorVariable.make(
         activation_checkpoint_op,
         **options,
     )
